SRM: don't try to rename non-existent tables
SRM db code manages the schema; if it detects the table has the 
wrong schema then it moves the old table away (currently by deleting 
it).  This allows the code to create the schema afresh a little 
later.

Unfortunately, the code to detect whether a table has the wrong 
schema doesn't distinguish between the table not existing (e.g., due 
to a fresh dCache install) or the table existing but having the 
wrong structure.  This results in unnecessary and confusing messages 
being logged when first starting up dCache, for example:

27 Sep 2011 18:09:57 (SRM-sisyphus) [] executing statement: DROP TABLE IF EXISTS putrequests_protocols CASCADE
27 Sep 2011 18:09:58 (SRM-sisyphus) [] executing statement: DROP TABLE IF EXISTS getrequests_protocols CASCADE

This patch fixes this by altering the behaviour so, if the table 
doesn't exist, SRM doesn't try to delete it.

Target: trunk
Patch: http://rb.dcache.org/r/3760/
Acked-by: Dmitry Litvintsev