cells: fix Formats.replaceKeywords
The cells package includes the Formats utility class.  This includes 
the replaceKeywords method that takes a String and replaces any 
reference to a keyword (e.g., "${keyword}") with some corresponding 
value.  In addition to CellShell, this method is also used in the new 
dCache configuration and in gPlazma2.

The current implementation contains two bugs: first, if a dollar is not 
the start of a reference then the following character is lost ("$shell" 
becomes "$hell").  Second, if the original string concludes with 
something that might be a reference then that string is lost (e.g., 
"foo$" --> "foo", "shell${foo" -> "shell").

This patch fixes these two bugs.  I'm concerned that, by fixing these 
problems, dCache's behaviour may be altered; in particular, if a site 
has implemented work-arounds (e.g., "$$this" to represent "$this") then 
this patch will break those configurations.  Because of this, only 
requesting this patch for trunk.

FOR RELEASE NOTES:

This patch fixes two related bugs in how dCache configuration is read.  
Previously, a dollar symbol that isn't part of a reference would 
swallow the following character ("$123" would be read as "$23").  The 
other bug was that, if line that ends with a dollar symbol or a 
potential reference then everything from the dollar symbol to the end 
of that line is lost ("a line ending with a dollar$" would be read as 
"a line ending with a dollar" and the line "another${line" would be 
read as "another")

A work-around for this bug is to double-up dollar symbols.  If the 
desired configuration is "$123" then the configuration has two dollar 
symbols "$$123".  When reading this, the second dollar symbol is 
swallowed and then intended output "$123" is provided.

With this release of dCache, non-reference dollars behave like any 
other character.  A "$123" in the configuration will be read as "$123".  
This may break a site's configuration if they rely on the double-dollar 
work-around described above.

Target: trunk
Patch: http://rb.dcache.org/r/3623/
Acked-by: Tanja Baranova
Acked-by: Gerd Behrmann