lm-server: don't print stack-trace on shutdown
lm-server has a thread that listens on UDP socket for incoming 
requests.  To interrupt this thread, the socket must be closed.  Doing 
this results in an exception being thrown, which is caught and logged, 
will a stack-trace.

This patch fixes this by ignoring this exception if the cell is 
shutting down.

Target: trunk
Patch: http://rb.dcache.org/r/3775/
Acked-by: Jan Schaefer