build: fix FHS server build
Patch http://rb.dcache.org/r/3665/ was committed to trunk as commit 
16203.  This patch adjusted the FHS packaging so that it included two 
additional files: chimera-get-acl and chimera-set-acl.  These two files 
were removed from trunk (after 1.9.12 branch but before the 1.9.13 
branch) so that, for 1.9.13 and trunk, these files are no longer 
available to be packaged.

Although the patch was specific to 1.9.12 branch it was committed to 
trunk and back-ported.  This allowed 1.9.12 to build FHS-compliant 
RPM-package builds but broke trunk in the process.

This patch fixes the problem by reverting c.16203.

By luck, the patch was not back-ported to 1.9.13, so that branch still 
builds FHS-compliant RPM packages.

Unfortunately, this problem was discovered after the 2.0-branch, so we 
need to fix that branch, too.

Target: trunk
Request: 2.0
Patch: http://rb.dcache.org/r/3759/
Acked-by: Christian Bernardt