UserShellAdmin: introducing new command  "repinfoof <pnfsid>"
Currently sys-admins need to execute 5 commands to get the information
of a state (for example cached or precious etc) of a file on all pools
where it stored. With this patch one can do it with only one command.

Ticket: RT#6688
(part of the feature request)

Acked-By: Tigran Mkrtchyan
Patch: http://rb.dcache.org/r/3777/