configuration: add 'immutable' as an annotation
Some property are declared to provide a standard location for similar 
information.  A possible example are the db.* properties (these allow 
the support scripts to provide database intervention operations in a 
service-agnostic fashion), although this isn't entirely clear, so 
this patch does not add the immutable flag at this time.

Such properties are not meant for the user to adjust; in fact, 
adjusting their values may lead to confusing, inconsistent behaviour.  
Despite this, we have no mechanism to enforce this.

This patch fixes this by introducing the 'immutable' annotation.  A 
property that is declared with this annotation may not be adjusted 
subsequently.  An attempt to do this will result in an error message 
being printed and dCache refusing to start.

Target: trunk
Patch: http://rb.dcache.org/r/3804/
Acked-by: Gerd Behrmann