admin: Upgrade to JLine 1.0
This patch upgrades the JLine library to version 1.0. JLine is not
available in Maven, which is why the patch places the library in
modules/external/ (just like the previous version).

A number of fixes have been made in version 1.0, including support for
proper support for lines that wrap. Delete key behaviour has also been
fixed. Command search (ctrl-R) has also been addded.

A number of fixes and workarounds had to be made in dCache to support
the new version of JLine. In particular the SSH input stream needs to
support the available() method, and I had to subclass the UnixTerminal
class (UnixTerminal assumes the input happens on a local TTY and that's
not the case for the admin service).

Since the patch resolves line wrapping, I propose to merge it into
stable branches.

Target: trunk
Request: 2.0
Request: 1.9.13
Request: 1.9.12
Acked-by: Karsten Schwank <karsten.schwank@desy.de>
Patch: http://rb.dcache.org/r/3791/