properties: tidy up obsolete and forbidden messages
When we declare a property as forbidden or obsolete the property's 
value is used to prove a meaningful message; if the property has no 
value then a default message is used.

Some of our properties have been declared as forbidden or obsolete but 
their values have been left as the default value.  This will lead to a 
cryptic or confusing message being reported back to the sys-admin.

This patch adjusts the values to provide meaningful messages.  The 
patch removes some properties' values when the message provides no 
additional information beyond the default message.  It also adjusts 
one property from obsolete to forbidden.

Patch: http://rb.dcache.org/r/3820/
Target: root
Request: 2.0
Request: 1.9.13
Request: 1.9.12
Acked-by: Jan Schaefer
Acked-by: Gerd Behrmann