parameters: add support for parametrised expansion
Our current configuration parameters support property value expansions 
by including a reference to a property's value.  The current 
implementation expects this reference to be a string-literal.

This patch adds support for parametrised references; that is, the name 
of a property may contain references to some other property; for 
example:

For example with the following property definitions:

    #  Valid values are 'first' and 'second'
    option = first

    foo-first = bar
    foo-second = baz

    foo = ${foo-${option}}

The reference "${foo}" will expand to the string "bar" by default (or 
if property 'option' is redefined with value "first") or "${foo}" will 
expand to "baz" if property 'option' is defined to have value "second".

Target: trunk
Patch: http://rb.dcache.org/r/3834
Acked-by: Gerd Behrmann
Acked-by: Jan Schaefer