srm: make srmLs more robust
The srmLs command requires a list of SURLs to list.  The existing code 
checks whether the list is null and returns the appropriate error if 
so.

Unfortunately, the test whether SURLs have been supplied is 
insufficient.  It's possible for a client to send a request where 
request.getArrayOfSURLs() returns a non-null object, but the 
getUrlArray() call returns null.  Such a request results in a 
NullPointerException.  This is somewhat ironically due to extracting 
the list of SURLs as a variable that is subsequently not used; 
however, even without this, creating the LsRequest fails.

This patch fixes this by adding more robust tests for malformed 
requests (ones without any SURLs).  The patch also provides a 
(hopefully) more meaningful error message and removes the unused 
variables.

Target: trunk
Request: 2.0
Request: 1.9.13
Request: 1.9.12
Request: 1.9.5
Patch: http://rb.dcache.org/r/3819/
Acked-by: Jan Schaefer
Acked-by: Dmitry Litvintsev