boot: Add plugin directory to dCache
This is the first in a series of patches to add a plugin structure to
dCache.

Plugins are represented as directories containing jar files, properties
files (containing defaults), and batch files (providing service
definitions).

The default plugin directory is /opt/d-cache/plugins for the OPT
packages and /usr/share/dcache/plugins:/usr/local/share/dcache/plugins
for the FHS packages (the first for packaged plugins and the second for
manually installed plugins). Plugins will be subdirectories of these
directories.

Eg the ALICE xrootd authorization plugin would be packaged as:

xrootd-authz-plugin-alice-1.0/
	README
	xrootd-authz-plugin-alice-1.0.jar
	xrootd-authz-plugin-alice-1.0.properties

This patch only adds the support for adding the JAR files to the
classpath. Subsequent patches will add support for properties and batch
files.

Question: Both 1.9.12 and 2.0 are marketed as having pluggable
components (gPlazma and pool manager), yet there is no official way to
load such plugins. Should this and subsequent patches be merged to fix
this?

Target: trunk
Request: 2.0
Request: 1.9.13
Request: 1.9.12
Patch: http://rb.dcache.org/r/3821/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>