chimera: Make delete permission check RFC 5661 compliant
RFC 5661 specifies two attributes to control file permission:
DELETE_CHILD and DELETE. The spec says that file deletion is permitted
if either attributes allow it to happen, if even the other attribute
denies the operation.  It also says that in case both are undefined,
then the behaviour is controlled by the ADD_FILE attribute.

Our currently implementation completely ignores this aspect of the RFC.

Patch: http://rb.dcache.org/r/3838/