admin: Add rudimentary tab completion
JLine supports tab completion plugins. This patch adds basic tab
completion based on the output of the help command. The help parser is
very basic and is fooled by some of our help strings. I will likely
replace it with a more complete parser written in Parboiled4j, but the
patch is functional and pretty useful as is.

Patch: http://rb.dcache.org/r/3815/