cells: fix race condition in getcellinfos
The getcellinfos command in CellShell provides a list of CellInfo 
objects for all the cells running in a domain or all those in the final 
stages of being cleaned up.  The command first makes a list of the 
names of the cells running in the domain.  It then fetches the CellInfo 
for each cell in the domain.

Unfortunately, the current implementation introduces a race-condition.  
If a cell finishes cleaning up after the list of cell names is obtained 
but before its CellInfo is obtained then a null is stored instead.

This patch fixes the code so any null entries are skipped.

Target: trunk
Request: 2.0
Request: 1.9.13
Request: 1.9.12
Request: 1.9.5
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6648
Patch: http://rb.dcache.org/r/3841/
Acked-by: Gerd Bahrmann