configuration: allow configuration of inter-domain cells messaging port
Current default is for dCache to use the cells package for 
inter-domain communication.  The default configuration is a 
star-topology, where the broker domain (dCacheDomain by default) 
receives TCP connections from all other domains.  The current 
configuration does not specify on which port the broker domain should 
listen for incoming connections.  The result is that a random port is 
chosen at run-time.

TCP connections to a random port number makes establishing firewall 
rules difficult.  It also makes it harder to understand is the 
responsibility of the different ports.

This patch adds a configuration option, broker.messaging.port, that 
controls on which port the broker domain will listen for incoming 
connections.  The default value is whatever port number the topology 
discovery service is configured: broker.port.

Target: trunk