scripts: add 'ports' command
This patch adds the 'ports' command to our dcache script.  This command 
lists which TCP and UDP ports dCache will listen on for this particular 
node, taking into account the node's configuration.

For some ports, whether dCache listens on them is determined at 
run-time (e.g., port-mapper for NFS).  Such ports are printed in 
parentheses.  Port ranges (as used by pools) are indicated with two 
numbers with a dash between them.

The mapping between dCache configuration and the ports dCache will 
listen on is stored as the property 'net.ports.tcp' and 
'net.ports.tcp'.  This property is defined in the default configuration 
for each service that listens on TCP ports.  The immutable annotation 
is used to prevent site-local configuration from updating these 
properties as such configuration would simply add confusion.

Broker communication is included, except for the UDP socket that 
non-broker domains open for topology discovery.  This is because the 
current behaviour is unconditionally to listen on a random port.  A 
future patch should allow configuration of this UDP port and add this 
UDP port to those listed by the ports command.

Target: trunk
Patch: http://rb.dcache.org/r/3793/
Acked-by: Gerd Behrmann
Acked-by: Jan Schaefer