cells: make lm client's listening UDP port configurable
In the cells package, lm provides a topology discovery / establishing 
service for cells inter-domain communication.  Currently, the lm client 
always listens on a randomly chosen UDP port for messages containing 
the topology information.

Using a random UDP port has the benefit that it's a static 
configuration that doesn't suffer from multiple domains attempting to 
acquire the same UDP port.

However, a site may choose to specify to which UDP port each domain 
should use.  This would allow tighter firewall rules, so providing some 
additional security.

This patch adds the ability to configure the UDP port to which the lm 
client will bind.  It is configured by the 'broker.client.port' 
property.  The allowed values are an integer or a literal 'random'.  
The former results in lm-client attempting to bind to the specified UDP 
port while the latter results in lm-client choosing a UDP port at 
random.

The default value for this property is 'random'.  This ensures 
backwards compatibility with existing deployments.

This patch also updates the ports command to reflect the lm-client UDP 
port.  After the patch, the ports command will list the UDP port used 
by a domain's lm-client.

Target: trunk
Patch: http://rb.dcache.org/r/3847/
Acked-by: Tigran Mkrtchyan
Acked-by: Jan Schaefer