documentation: remove README-GLUE
The file README-GLUE is supplied with dCache.  It contains information 
about configuring and diagnosing problems with the GLUE info-provider.

The file is now redundant: this information (and more) is available in 
the dCache Book; also, the information in README-GLUE isn't modified 
with the FHS layout; therefore, it provides incorrect information for 
the FHS builds.

This patch simply removes the file.

Target: trunk
Patch: http://rb.dcache.org/r/3861/
Acked-by: Karsten Schwank