cells: add hasOption method to Args
The cells jar file contains an arguments parsing class Args.  This 
parses a command line and provides this information for code that 
depends on this command line.

There are many places in dCache code-base that has code like:

    args.getOpt("foo") != null

This is a somewhat obscure way of testing whether the option "foo" has 
been specified (".. -foo ..").  The expression evaluates to true 
whether the option has any argument part ("-foo" or "-foo=bar").  The 
negated form is also used, albeit less frequently:

    args.getOpt("foo") == null

Such tests are undesirable because:

    o  the code's intention isn't obvious (what does "null" mean?
       Does it include the case when an option is specified by no
       argument given?),

    o  it relies on the behaviour of getOpt when there is no option,

    o  it uses null as a special case,

    o  it may require parentheses to ensure correct functionality when used
       with higher precedence operators, so making code harder to read and
       error-prone,

    o  reading the code requires a mental-map to figure out whether the code
       tests the presence or absence of an option, so is error-prone.

This patch adds a hasOption method to the Args class.  This is a 
boolean method that takes the name of an option and returns true iff 
the option has been specified in the command line.  The patch updates 
the unit tests to verify hasOption provides the expected results.

The patch also adjusts dCache code to make use of this new method.  
This part of the patch was achieved by searching through the code-base 
and adjusting it automatically with a short script.  Here is this 
script:

    #!/bin/bash

    for dir in modules/cells modules/dCache modules/dCacheJUnit modules/srm; do

        #  Map  "getOpt(..) != null" or "getOption(..) != null" to  "hasOption(..)"
        find $dir -name \*.java | xargs sed -ie "s/getOpt\(ion\)\? *(\(\([^()]*\|[^()]*([^()]*)[^()]*\)\)) *!= *null/hasOption(\2)/g"

        #  Map  "foo.getOpt(..) == null" or "foo.getOption(..) == null"  to
        #  "!foo.hasOption(..)"
        find $dir -name \*.java | xargs sed -ie 's/\([a-zA-Z_][a-zA-Z0-9_]*\)\.getOpt\(ion\)\? *(\([^)]*\)) *== *null/!\1\.hasOption(\3)/g'

        #  Remove unnecessary parentheses for some easy cases (when statement is
        #  an assignment, a return value or part of a boolean expression)
        find $dir -name \*.java |xargs sed -ie 's/\(\(=\|return\|&&\|||\) *\)( *\(!\)\? *\([a-zA-Z_][a-zA-Z0-9_]*\.hasOption([^)]*)\) *)/\1\3\4/g'

        #  Remove backup files generated by sed
        find $dir -iname \*.javae -exec rm \{\} \;
    done

Note that the RE "(\(\([^()]*\|[^()]*([^()]*)[^()]*\)\))" matches the 
arguments to the getOpt or getOption methods (including the 
parentheses) with the argument being either something without 
parentheses (a literal value, variable, etc) or something with a single 
pair of parentheses (a simple method call).  Although this doesn't 
cover all possibilities, it is sufficiently general for dCache's 
code-base.

As a nice bonus, the converted lines are shorter.

This patch does not remove all cases where getOpt return value is 
tested against null.  Most notably, there are several places where 
there is code of the form:

    String foo;

    if((foo = args.getOpt("bar")) != null) {
        // code that uses variable foo
    }

In addition, there may be other places where the result of getOpt 
method is saved in a variable and this variable is later compared 
against null.

Target: trunk
Patch: http://rb.dcache.org/r/3855/
Acked-by: Tigran Mkrtchyan