info-provider: fix the GLUE2 we publish
The info-provider can publish Glue v1.3, v2.0 or both.  By default it 
publishes both.

Glue (both versions) include the concept of objects that are linked 
together.  Recently, how some of the links are represented in Glue 2 
has changed.  This patch brings our info-provider in line with the new 
names of these links.  The patch also fixes a number of minor problems 
that are also present in the current implementation.

One of these problems is with how we represent which VOs may 
read/write/stage into pools that are only accessible via *@* storage 
class.  If the site has configured their info-provider so that no VOs 
are associated with *@* then the info-provider tries to publish an 
object when it shouldn't.  However, this object has a broken 
structure.  The broken structure means that OpenLdap will reject this 
object, so achieving the desired result.  However, some warning 
messages are printed into the BDII log file.

Because of limitations in xylophone, this problem is currently not 
easy to solve.  A future patch will address this problem; however, 
this patch documents the problem with a FIXME.

FOR RELEASE NOTES:

Be sure you have at least v2.0.8 of glue-schema RPM installed on the 
node running the info-provider.

Target: trunk
Request: 2.0
Request: 1.9.13
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6633
Patch: http://rb.dcache.org/r/3880/
Acked-by: Tigran Mkrtchyan