defaults: update description of obscure feature
dCache uses a star topology for cells communication: dCacheDomain 
accepts connections from any/all other domains and routes messages 
accordingly.  However, the cells package supports more complicated 
topologies and dCache may be configured to use alternative topologies 
by writing a configuration file.  The default location of the file is 
in the etc directory; however, this disables the 'save' option within 
the lmd cell, preventing on-line adjustment of the topology that the 
sysadmin wishes to save.

We received a ticket that the file should be more correctly placed 
somewhere in /var, much in keeping with the pool-manager file.

In the ticket, I argued that, since we neither document the file format 
of this file (or the admin interface for adjusting it), nor (to the 
best of my knowledge) is this feature used anywhere, perhaps a more 
robust response would be to simply remove this option.

This patch is a compromise between these two positions: it keeps the 
option as-is, but adds some additional information that may help 
prevent confusion in the future.

Target: trunk
Request: 2.0
Request: 1.9.13
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6746
Patch: http://rb.dcache.org/r/3892/
Acked-by: Karsten Schwank