configuration: allow escaping of characters
The configuration system has properties that allow for references to 
other properties that are expanded when their values are needed using a 
simple "${name}" syntax, where "name" is the name of some property 
value.  It also allows for embedded references, so "${${foo}}" will 
take the value of "foo" and use it as the name of a property that is 
then expanded; for example if property "foo" has value "bar" and 
property "bar" has value "baz" then "${${foo}}" expands to "baz".

One limitation of the current code is that there is no provision for 
escaping characters: for enforcing that a character is taken as a 
literal value.  If a property reference can be expanded then it is 
always replaced with that property's value.

This patch adds support for literal values.  A backslash character 
('\') ensures that the following character is taken literally and 
doesn't result in any expansion.

A backslash character may precede any character.  The character 
following a backslash character is treated as a literal character for 
that expansion.  Escaping a normal character (one that is not special) 
is allowed and the two-character combination is replaced with the 
second character ("\foo" --> "foo").

If the character appears at the end of the property's value then it is 
taken as a literal backslash ("foo\" --> "foo\")

Since the backslash character is now a special character, any reference 
to it must be escaped ("foo\\bar" --> "foo\bar")

If a property has a value "${foo}" then there are seven ways of 
preventing this string from being expanded: escaping the '$' (as in 
"\${foo}"), escaping the '{' (as in "$\{foo}"), escaping the '}' 
("${foo\}") and the four combinations of these three possibilities 
("\$\{foo}", etc).  The ways of preventing expansion that escape the 
closing brace work because the expansion becomes incomplete, resulting 
in the literal value being used.  However, escaping the closing brace 
will affect how any future expansions are treated, which may confuse 
the unwary, so isn't recommended.

With escaping, property names may now contain '$', '{', '}' characters.  
However, references to these properties should be escaped twice, 
although for some property names, a single level of escaping will be 
sufficient.

For simple names that include '$', '{' or '}' but not in a fashion that 
represents a further expansion then single escaping will work.

For example, if a property has name "${foo" and value "bar" then 
"${\${foo}", "${$\{foo}" and "${\$\{foo}" will all expand to 'bar'.  If 
there is a property "}foo" with value "bar" then "${\}foo}" will expand 
to "bar".

Properties with names that include a '\' character or that look like a 
property expansion must be escaped twice; e.g., a reference to the 
value of property with name "foo\bar" must be written "${foo\\\\bar}".

Property expansions that make use of embedded references are also 
supported; but an additional level of escaping is needed.  For 'simple' 
properties, this results in double-expanding.

For example, if there is a property called "}foo" with value "bar" and 
a property "bar" with value "baz" then "${\}foo}" will expand to "bar" 
(as above) and "${${\\\}foo}}" will expand to "baz".

If a property expansion that makes use of an embedded reference to a 
property with '/' then three levels of escaping are needed.

For example, if a property called "foo\bar" has value "baz" and a 
property "baz" has value "fruitbat" then "${foo\\\\bar}" has value 
"baz" (as above) and "${${foo\\\\\\\\bar}}" has value "fruitbat".

Although it works, it is not recommended to use '$', '{', '}' and '/' 
in property names; this is especially so for '/'.

Target: trunk
Patch: http://rb.dcache.org/r/3895/
Acked-by: Gerd Behrmann