Make FHS and OPT output directory
This way you can specify where all the output of a mvn build shall be dumped. This is mainly important for building RPMS and alike.

The directory does not seem to configurable just using -DoutputDirectory as can be seen in the doc:

http://maven.apache.org/plugins/maven-assembly-plugin/single-mojo.html

Patch:http://rb.dcache.org/r/3926/

Acked-by:Tigran Mkrtchyan
Target:trunk