info: add dga service to decouple info service from creation of DGA
Info has the concept of data-gathering activity.  These are scheduled 
activities where the object decides when it should be next run, what to 
do when invoked and takes responsibility for processing and reply 
messages.  A DGA has access to the existing state, so can base its 
decisions (what to do now, when next to request running) on the state 
gathered so far.

The current code creates all DGAs within the info-service.  This 
creates a tight coupling between different services within dCache: the 
info service needs to know what messages to send to obtain the 
information it needs.  As we wish to move components into separate 
packages (as part of the maven-isation process), this tight coupling 
becomes a problem.

To solve this problem, the info service uses Java's ServiceLoader class 
to discover all the DgaFactoryService implementations at run time.  
Each DgaFactoryService can create one or more DGAs that are injected 
into the scheduler.

The patch also moves the existing monitoring activity into separate 
packages, with each package providing an independent DgaFactoryService.  
In addition to tidying up the code, this demonstrates that the the 
mechanism works as expected.

Target: trunk
Patch: http://rb.dcache.org/r/3904/
Acked-by: Gerd Behrmann