properties: allow escaping of property references
The configuration system has properties that allow for references to 
other properties that are expanded when their values are needed using a 
simple "${name}" syntax, where "name" is the name of some property 
value.  It also allows for embedded references, so "${${foo}}" will 
take the value of "foo" and use it as the name of a property that is 
then expanded; for example if property "foo" has value "bar" and 
property "bar" has value "baz" then "${${foo}}" expands to "baz".

One limitation of the current code is that there is no provision for 
escaping characters: for enforcing that a character is taken as a 
literal value.  If a property reference can be expanded then it is 
always replaced with that property's value.

This patch adds support for literal dollar symbols.  This allows a 
property value to include write a dollar symbol without that dollar 
becoming part of a potential expansion.  Such a literal dollar is 
written as a two-dollar sequence ("$$").

If a property foo should have the value "${bar}" (as a literal sequence 
of characters) then declaring property foo has value "$${foo}" will 
give the desired result: references to foo ("${foo}") will expand to 
"${bar}" irrespective of whether a property bar is defined.

As a side-effect of this patch, property names may now include a '${' 
sequence.  Such properties may be referenced by placing an escaped 
version of the name in the braces; for example, if property "${foo" has 
value "bar" then the reference "${$${foo}" will expand to "bar".

Target: trunk
Patch: http://rb.dcache.org/r/3902/
Acked-by: Karsten Schwank