common: Move Subjects from cell and dcache module to common
The motivation for the patch is that the our principals have been in 
common for a while and thus Subjects belongs there too. This is made 
clear by cells having a "mini Subjects" class to workaround the 
placement of Subjects in dcache.

The change is also necessary to split out the acl code and chimera code 
from dCache: That code depends on the principals and on the Subjects 
class.

Unit tests pass as do my transfer tests.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de> (per email)
Patch: Not on RB as RB could not handle the patch