chimera: Reduce contention point on file deletion
When a file is deleted, the nlink of the parent directory is
decremented.

During bulk deletion of files in the same directory, this update of the
parent inode causes serialization of the transactions. Once one
transaction has issued an update on the parent inode, other transactions
will block on their update until the first transaction is committed.

This problem is a major bottleneck at NDGF in which we often see
transfers fail during bulk deletion.

This patch tries to address the issue by reducing the time between the
update of the parent inode and the commit of the transaction. It
achieves this by moving the decrease of the nlink and the update of the
level-0 mtime to the last steps performed during file removal. In
particular those updates are now done after the file inode itself is
moved to the trash table (that happens with an SQL trigger).

Target: trunk
Patch: http://rb.dcache.org/r/3947/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>