webdav: Fix buffering issues in the door
Milton uses a BufferingOutputStream wrapper around the servlet's
OutputStream. It uses this to determine the content length if the
resource doesn't know the length ahead of time. It also uses the
BufferingOutputStream is the buffering property of the response handler
isn't set (which was the case for dCache). As a consequence
BufferingOutputStream was used for both file downloads proxied by the
door and for directory listings.

BufferingOutputStream uses an in-memory buffer for the first 100KB and
falls back to using a temporary file for larger content. Besides
introducing latency and causing IO on the system disk of the door,
Milton doesn't always deal nicely with error situations. If something
fails it relies on Java object finalization to remove the temporary
file, which in turn may lead to a large number of temporary files to
accumulate before they get removed.

The buffering is however completely unnecessary: For files the content
length is known and no buffering is required. For directories the
content length isn't known, but Jetty deals nicely with that situation.
Jetty will use chunked encoding for HTTP/1.1 and for HTTP/1.0 it falls
back to closing the stream to indicate the end of the content.

The patch disables Milton's buffering logic. It does this by setting the
buffering property on the response handler to always (which is a rather
unintuitive value for this behaviour). The patch also works around
Milton insisting on setting an empty value for Content-Length if the
length is unknown. Instead we should refrain from adding a
Content-Length header at all if the content length is not known.

Target: trunk
Request: 2.0
Request: 1.9.12
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6808
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6716
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Jan Schaefer <jan.schaefer@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/3972/