httpd: Pool SelectionUnit Page is not reading HttpRequest Parameters - fixing this
The old implementation does not read the Parameters of the Http-GET Request. I guess that this broke with the introduction of jetty for this service. This leads to a non-working match functionality and a non-working (though unknown to me) grep and sort functionallity of TapeTransferPage. As there have been no tickets on this issue it doesn't seem that these are widely used features considering that this bug is in there approx 1,5 years if it worked ever...

patch:http://rb.dcache.org/r/3983/

acked-by:

Target: trunk
Request: 2.0
Request: 1.9.12