scripts: check whether hostkey.pem is stored in PKCS#8 format
Starting with OpenSSL v1.0 (or there abouts) the 'pkcs12' utility 
(commands that starts 'openssl pkcs12') will write private keys in 
PKCS#8 format.  Unfortunately, jGlobus 1.8 requires the certificate to 
be in PKCS#1 format.  With earlier versions of OpenSSL, the pkcs12 
utility used the PKCS#1 format for private keys.

The pkcs12 utility is commonly used to convert the file exported by web 
browsers [that contain both certificate(s) and private key(s)] into 
separate PEM-encoded certificate and private key files.  It is commonly 
documented as how to convert host certificates requested via a 
web-browser.

It is also possible that OpenSSL v1.0 (+/- \delta) when generating 
private keys with a signing request, saves the private key in PKCS#8 
format.

The OpenSSL 'rsa' command still writes private keys in PKCS#1 format 
and is able to read PKCS#8-formatted private keys. Therefore a simple 
work-around is to run a PKCS#8-formatted private key through a 'openssl 
rsa' command that only alters the format.

This patch adds a check to all dcache commands to check whether the 
hostkey.pem is in PKCS#8 format, if present.  If so, a warning is 
emitted and the command returns with an error code.  The warning 
includes instructions describing how to fix the problem.

NB. The patch doesn't apply to v1.9.5 so a separate patch will be 
submitted for that branch.

Target: trunk
Request: 2.0
Request: 1.9.12
Patch: http://rb.dcache.org/r/3985/
Acked-by: Tigran Mkrtchyan