nfs: allow reuse of listening port
A shutdown of the NFS server (located on the nfs door and part of the 
pools) can leave the TCP connections in TIME_WAIT state.  This happens 
if clients are connected when the component is stopped: if the clients 
disconnect first then there is no problem.  Normally, the Linux server 
refuses to allow a server to bind to a port if there are TCP 
connections in TIME_WAIT state.  The option SO_REUSEADDR switches off 
this behaviour, allowing software to bind to a port that has 
connections in TIME_WAIT state.

The NFS server currently doesn't initiate connections with the 
SO_REUSEADDR option set.  Because of this, doors or pools may find 
themselves unable to the port they used before.  This problem is 
particularly acute for doors since they must use port 2049.  This can 
lead to a dCache nfs-door or pool failing to restart if that component 
was connected when the restart was initiated.

This patch alters the NFS server so the SO_REUSEADDR option is set.

Target: trunk
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6839
Request: 2.0
Request: 1.9.12
Request: 1.9.5
Patch: http://rb.dcache.org/r/4007
Acked-by: Dmitry Litvintsev
Acked-by: Tigran Mkrtchyan