gPlazma1 - fix UnsupportedOperationException when using Kerberos
I noticed the following bug:

02 Dec 2011 15:01:56 (gPlazma) [] Uncaught exception in thread pool-6-thread-18
java.lang.UnsupportedOperationException: null
        at java.util.AbstractList.add(AbstractList.java:148)
        at java.util.AbstractList.add(AbstractList.java:108)
        at gplazma.authz.AuthorizationController.authorize(AuthorizationController.java:192)
        at org.dcache.auth.GplazmaLoginStrategy.login(GplazmaLoginStrategy.java:109)
        at org.dcache.services.login.MessageHandler.messageArrived(MessageHandler.java:113)
        [...]

This happens when attempting to use gPlazma-1 and logging in with a 
Kerberos token; in my case, this was with NFS 4.1/pnfs + 
Kerberos-based authentication.

The problem is in the gPlazma-1 code.

It looks at the roles list and, if it is empty, adds an additional 
place-holder value so that when it iterates over the list of plugins 
it tries will a 'null' instead of a specific role.  When logging in 
with a Kerberos token there are never any roles.  In this particular 
case, the dCache-gPlazma1 bridge code uses a Collections.emptyList.  
This breaks the gPlazma-1 'trick' since this collection doesn't allow 
insertion of any entries.

Target: trunk
Request: 2.0
Request: 1.9.12
Patch: http://rb.dcache.org/r/4006
Acked-by: Gerd Behrmann